+++ tests/test_layer_profiling/test_column_profiling/test_decimal_profiling.py
@@ -1,0 +1,9 @@
+from respark.layer_profile import (
+    DecimalColumnProfile,
+    profile_decimal_column,
+)
+
+
+def test_creates_valid_DecimalProfile():
+    assert type("add after creating Decimal test data") == str
+
