+++ tests/test_layer_profiling/test_column_profiling/test_fractional_profiling.py
@@ -1,0 +1,10 @@
+from respark.layer_profile import (
+    FractionalColumnProfile,
+    profile_fractional_column,
+)
+
+
+def test_creates_valid_FractionalProfile():
+
+    assert type("add after creating Fractional test data") == str
+
